writing tests for catalog.py and cleaning up the setup
